src/requirements.txt
langgraph
langchain
grandalf
typer
colorama
dotenv
openai
langchain_openai
GitPython
PyGithub
black